ARQ-138 Corrected the names for the Maven module and the Java package.
