--- arquillian-wls-remote-10.3.x/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,0 @@
-org.jboss.arquillian.container.wls.remote_103x.WebLogicExtension
