--- arquillian-wls-remote-10.3.x/src/test/resources/in-container-web-no-cdi.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app version="2.5"
-   xmlns="http://java.sun.com/xml/ns/javaee"
-   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-   xsi:schemaLocation="
-      http://java.sun.com/xml/ns/javaee
-      http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd">
-
-   <env-entry>
-      <env-entry-name>resourceInjectionTestName</env-entry-name>
-      <env-entry-type>java.lang.String</env-entry-type>
-      <env-entry-value>Hello World from an env-entry</env-entry-value>
-   </env-entry>
-
-   <servlet>
-      <servlet-name>TestServlet</servlet-name>
-      <servlet-class>org.jboss.arquillian.container.wls.remote_103x.MyServlet</servlet-class>
-   </servlet>
-
-   <servlet-mapping>
-      <servlet-name>TestServlet</servlet-name>
-      <url-pattern>/Test</url-pattern>
-   </servlet-mapping>
-
-</web-app>
