+++ wls-remote-10.3/src/main/java/org/jboss/arquillian/container/wls/remote_10_3/WebLogicExtension.java
@@ -1,0 +1,38 @@
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright 2011, Red Hat Middleware LLC, and individual contributors
+ * by the @authors tag. See the copyright.txt in the distribution for a
+ * full listing of individual contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jboss.arquillian.container.wls.remote_10_3;
+
+import org.jboss.arquillian.container.spi.client.container.DeployableContainer;
+import org.jboss.arquillian.container.test.spi.client.deployment.ProtocolArchiveProcessor;
+import org.jboss.arquillian.core.spi.LoadableExtension;
+
+/**
+ * The Arquillian integration for WLS 10.3.x.
+ * 
+ * @author Vineet Reynolds
+ *
+ */
+public class WebLogicExtension implements LoadableExtension
+{
+
+   public void register(ExtensionBuilder builder)
+   {
+      builder.service(DeployableContainer.class, WebLogicContainer.class)
+             .service(ProtocolArchiveProcessor.class, WebLogicCDIProcessor.class);
+   }
+
+}
