+++ wls-remote-10.3/src/test/resources/arquillian.xml
